logging added

+++ database/migrations/2026_03_29_134859_create_useractivity_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('useractivity', function (Blueprint $table) {
             $table->id();
-            $table->integer('userid');
+            $table->unsignedBigInteger('userid');
             $table->string('endpoint_used', 256);
             $table->integer('files_downloaded');
             $table->date('activity_date');
